requirements.txt
allure-pytest
pytest
selenium
webdriver-manager
pytest-xdist
pytest-rerunfailures
